+++ src/config/index.js
@@ -1,3 +1,4 @@
-module.exports ={
-    SUPPORTED_TOKENS_URL : 'https://raw.githubusercontent.com/getsafle/multichain-data/main/uniswap.json',
+module.exports = {
+    SUPPORTED_TOKENS_URL: 'https://raw.githubusercontent.com/getsafle/multichain-data/main/uniswap.json',
+    SUPPORTED_CHAINS: ['ethereum', 'polygon'],
 }